Incremental work on following tickets.

+++ src/Orchard.Web/Modules/OrchardPros/Helpers/EnumerableHelpers.cs
@@ -3,7 +3,7 @@ using OrchardPros.Models;
 
 namespace OrchardPros.Helpers {
     public static class EnumerableHelpers {
-        public static IPagedList<T> ToPagedList<T>(this IEnumerable<T> collection, int totalCount) {
+        public static IPagedList<T> ToPagedList<T>(this IEnumerable<T> collection, long totalCount) {
             return new PagedList<T>(collection, totalCount);
         }
     }
