+++ src/Orchard.Web/Modules/OrchardPros/Models/PagedList.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 namespace OrchardPros.Models {
     public class PagedList<T> : List<T>, IPagedList<T> {
-        public PagedList(IEnumerable<T> collection, int totalItemCount) : base(collection) {
+        public PagedList(IEnumerable<T> collection, long totalItemCount) : base(collection) {
             TotalItemCount = totalItemCount;
         }
